First samples converted
